requirements.txt
groq==0.6.0
langchain==0.2.0
langchain_groq==0.1.4
Pillow==10.0.0
Pillow==10.3.0
psutil==5.9.5
python-dotenv==1.0.1
pyttsx3==2.90
rembg==2.0.56
SpeechRecognition==3.10.0
streamlit==1.34.0
wikipedia==1.4.0
~treamlit==1.29.0
